chore: rename NFT to Nft
