+++ apps/web/src/components/Settings/Profile/NftPicture.tsx
